// frontend/src/lib/generated/SightingsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GroupedSighting } from "./GroupedSighting";
import type { Sighting } from "./Sighting";

export type SightingsResponse = { sightings: Array<Sighting> | null, groups: Array<GroupedSighting> | null, total: bigint, page: number, page_size: number, total_pages: number, };