// frontend/src/lib/generated/GroupedSighting.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type GroupedSighting = { common_name: string | null, scientific_name: string | null, country_code: string | null, trip_name: string | null, observed_at: string | null, count: bigint, };