// frontend/src/lib/generated/Sighting.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Sighting = { id: bigint, common_name: string, scientific_name: string | null, count: bigint | null, latitude: number, longitude: number, country_code: string | null, region_code: string | null, observed_at: string, };